added checks to see whether or not user is logged in before going to organization registration, settings, or installation

+++ controllers/molecules.js
@@ -6,32 +6,35 @@ var multer	 = require('multer');
 var fs			 = require('fs');
 
 router.post('/upload_mol', multer({ dest: './public/uploads' }).single('molecule'), (req, res, next) => {
-	var gfs 				= req.app.get('gridfs');
-	var writestream = gfs.createWriteStream({ filename : req.file.originalname});
-	fs.createReadStream(req.file.path).pipe(writestream).on('close', (file) => {
-		var file_id = file._id;
-		fs.unlinkSync(req.file.path);
-		new molecule({
-			_author  		: req.user._id,
-			title		 		: req.body.title,
-			category 		: req.body.category,
-			description : req.body.description,
-			molfile			: file_id
-		}).save((err, mol) => {
-			if (err) { res.render('dashboard', { message : 'error', isAuth : req.isAuthenticated() }); }
-			else {
-				user.findByIdAndUpdate(
-					req.user._id,
-					{$push: {'molecules': mol._id }},
-					{safe: true, upsert: true, new: true},
-					(err) => {
-						if (err) { res.render('dashboard', { message : 'error', 	 isAuth : req.isAuthenticated() }); }
-						else 		 { res.render('dashboard', { message : 'success', isAuth : req.isAuthenticated() }); }
-					}
-				);
-			}
+	if (!req.user) { res.render('login', 				{ message: 'You must login first!', isAuth : req.isAuthenticated() }); }
+	else {
+		var gfs 				= req.app.get('gridfs');
+		var writestream = gfs.createWriteStream({ filename : req.file.originalname});
+		fs.createReadStream(req.file.path).pipe(writestream).on('close', (file) => {
+			var file_id = file._id;
+			fs.unlinkSync(req.file.path);
+			new molecule({
+				_author  		: req.user._id,
+				title		 		: req.body.title,
+				category 		: req.body.category,
+				description : req.body.description,
+				molfile			: file_id
+			}).save((err, mol) => {
+				if (err) { res.render('dashboard', { message : 'error', isAuth : req.isAuthenticated() }); }
+				else {
+					user.findByIdAndUpdate(
+						req.user._id,
+						{$push: {'molecules': mol._id }},
+						{safe: true, upsert: true, new: true},
+						(err) => {
+							if (err) { res.render('dashboard', { message : 'error', 	 isAuth : req.isAuthenticated() }); }
+							else 		 { res.render('dashboard', { message : 'success', isAuth : req.isAuthenticated() }); }
+						}
+					);
+				}
+			});
 		});
-	});
+	}
 });
 
 module.exports = router;
